fix: update building steps
The use of the compile scripts was the old way and the compiler file provided was only to assist in backwards compatibility.

I updated the example to still compile to the name you wanted.  However, the preferred way is to use an `assembly` folder with `index.ts` and then the name of the contract comes from a package.json.