Update celo-hardhat-deploy-and-nft-app.md
Suggested changes to earlier portions of the tutorial. This is extremely code-heavy and needs a solid formatting and explanation pass before it can be made live. @lucasespinosa28 please reconsider use of subheadings to achieve bold text between code blocks. It would be better for the formatting and flow of this document to break the tutorial down into a few smaller sections and use appropriate 1st and 2nd level headings.

I have edited the document directly for this round of suggestions as to illustrate the kind of changes we're looking for. I am happy to work with you to fine tune this tutorial because it is quite good, it just needs a little bit more editing. Review my suggested edits and let me know what you think! 😄