Update referenced constant name to be accurate
On line 68 we reference three constants, `wsProvider`, `api` and `keyring`. In the code sample being referenced, `wsProvider` does not exist and is instead named `httpProvider`, this updates the reference on line 68 to match the code sample. (The name in the code sample also matches preceding code samples in previous steps.)