Update write-and-deploy-a-smart-contract-on-near.md
Three changes:

(1) I've replaced the call to `cargo init` by `cargo init --lib` as it make the overall set up shorter and clearer

(2) A step on contract name have been added on deploy section in order to clarify the subsequent call on `near-cli`, and it's a good practice to give an identifier to our contract.

(3) When we call a view method of our contract (one who doesn't mutate the state of our contract) we've to pass the view parameter to near-cli  in place to call. Otherwise, we'll have to go through a transaction, then it will takes some times an cost some fee.